Correction des merges + ajout indication action retrait

+++ src/main/java/fr/fges/menu/actions/RemoveGameEntry.java
@@ -1,8 +1,11 @@
 package fr.fges.menu.actions;
 import fr.fges.formatters.MenuInterface;
+import fr.fges.models.BoardGame;
 import fr.fges.models.commands.RemoveGameCommand;
 import fr.fges.repositories.GameCollectionDao;
 import fr.fges.repositories.HistoryDao;
+
+import java.util.Optional;
 
 public class RemoveGameEntry implements MenuEntry {
     private final String label;
@@ -16,11 +19,14 @@ public class RemoveGameEntry implements MenuEntry {
     @Override
     public void handle(MenuInterface UI, GameCollectionDao dao) {
         String title = UI.getGameTitle();
-        history.saveModification(new RemoveGameCommand(dao.findByTitle(title).get()));
+        Optional<BoardGame> game = dao.findByTitle(title);
+        if (game.isEmpty()) {
+            UI.displayMessage("No board game found with that title.");
+            return;
+        }
+        history.saveModification(new RemoveGameCommand(game.get()));
         if (dao.delete(title)) {
             UI.displayMessage("Board game removed successfully.");
-        } else {
-            UI.displayMessage("No board game found with that title.");
         }
     }
